Added modules, refactored some code
