+++ data_manager/__init__.py
@@ -1,0 +1,11 @@
+from .saves import SavesManager
+from .path_gen import Difficulty, path_gen
+
+# saves.py
+SavesManager = SavesManager
+
+# path_gen
+Difficulty = Difficulty
+path_gen = path_gen
+
+# TODO :: Change saves manager to use json/yaml (maybe encode/decode puzzle)
