+++ other_widgets/__init__.py
@@ -1,0 +1,3 @@
+from .stopwatch import StopwatchWidget
+
+CStopwatch = StopwatchWidget
