+++ page/__init__.py
@@ -1,0 +1,4 @@
+from page import game, settings
+
+gamePage = game
+settingsPage = settings
