+++ page/pages.py
@@ -1,0 +1,4 @@
+from data_manager import Difficulty, SavesManager, path_gen
+
+
+savesManager = SavesManager()
